requirements.txt
Flask>=2.0,<3.0
tensorflow==2.11.0
protobuf==3.19.6
jinja2==3.0.3
eventlet>=0.33.3
dnspython>=2.2.1
python-socketio==4.2.1
